+++ dev_deps.ts
@@ -1,3 +1,6 @@
 // Copyright 2021-present the Fonction authors. All rights reserved. MIT license.
-export { assert, assertEquals } from 'https://deno.land/std/testing/asserts.ts'
+export {
+  assert,
+  assertEquals
+} from 'https://deno.land/std@0.97.0/testing/asserts.ts'
 export { spy } from 'https://deno.land/x/mock@v0.9.5/spy.ts'