+++ test/isLength0.test.ts
@@ -1,0 +1,56 @@
+// Copyright 2021-present the Fonction authors. All rights reserved. MIT license.
+import { assertEquals } from '../dev_deps.ts'
+import { isLength0 } from '../src/isLength0.ts'
+import { isSymbol } from '../src/isSymbol.ts'
+import {
+  BIG1,
+  DATE,
+  EMPTY_ARRAY,
+  EMPTY_OBJECT,
+  EMPTY_STRING,
+  MAP,
+  ONE,
+  SET,
+  SYMBOL,
+  VOID_FN,
+  VOID_PROMISE,
+  WEAK_MAP,
+  WEAK_SET,
+  ZERO
+} from './index.ts'
+
+Deno.test('isLength0', () => {
+  const table: [unknown, boolean][] = [
+    [EMPTY_STRING, true],
+    [EMPTY_ARRAY, true],
+    [{ length: 0 }, false],
+    [ZERO, false],
+    [ONE, false],
+    [NaN, false],
+    [Infinity, false],
+    ['test', false],
+    [false, false],
+    [true, false],
+    [BIG1, false],
+    [SYMBOL, false],
+    [null, false],
+    [undefined, false],
+    [EMPTY_OBJECT, false],
+    [{ nest: {} }, false],
+    [[[]], false],
+    [MAP, false],
+    [SET, false],
+    [WEAK_MAP, false],
+    [WEAK_SET, false],
+    [VOID_FN, false],
+    [VOID_PROMISE, false],
+    [DATE, false]
+  ]
+  table.forEach(([val, expected]) => {
+    assertEquals(
+      isLength0(val),
+      expected,
+      `isLength0(${isSymbol(val) ? 'symbol' : val}) -> ${expected}`
+    )
+  })
+})